change var mu from float to double
git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/undo@2061 3b3fb362-3133-0410-aa15-cf69e0a59cb7

+++ src/core/include/hydrogen/sampler/Sampler.h
@@ -94,7 +94,7 @@ private:
 	unsigned __render_note( Note* pNote, unsigned nBufferSize, Song* pSong );
 
         InterpolateMode __interpolateMode;
-        float Interpolate( float y0, float y1, float y2, float y3, float mu );
+        float Interpolate( float y0, float y1, float y2, float y3, double mu );
 
 	int __render_note_no_resample(
 	    Sample *pSample,
